feat: wrap up initial fixtures for streaming LLM
