--- apps/nextjs/src/app/api/chat/fixtures/recordings/chat-1.formatted.txt
@@ -1,79 +1,0 @@
-{
-  "type": "llmMessage",
-  "patches": [
-    {
-      "type": "patch",
-      "reasoning": "These prior knowledge points ensure that pupils have the foundational understanding necessary for this lesson.",
-      "value": {
-        "type": "string-array",
-        "op": "add",
-        "path": "/priorKnowledge",
-        "value": [
-          "The Romans ruled Britain for several centuries",
-          "Roman Britain had roads, towns, and baths",
-          "The Romans built Hadrian's Wall to protect against northern tribes"
-        ]
-      }
-    },
-    {
-      "type": "patch",
-      "reasoning": "These key learning points highlight the main aspects pupils should understand by the end of the lesson.",
-      "value": {
-        "type": "string-array",
-        "op": "add",
-        "path": "/keyLearningPoints",
-        "value": [
-          "The Romans left Britain around 410 AD",
-          "Roman withdrawal led to social and economic changes in Britain",
-          "Post-Roman Britain saw the rise of new groups such as the Anglo-Saxons"
-        ]
-      }
-    },
-    {
-      "type": "patch",
-      "reasoning": "These common misconceptions may arise due to the complexity of historical events and timelines.",
-      "value": {
-        "type": "misconceptions",
-        "op": "add",
-        "path": "/misconceptions",
-        "value": [
-          {
-            "misconception": "The Romans left Britain suddenly",
-            "description": "Pupils might think the Romans left overnight. Explain that it was a gradual process due to pressures elsewhere in the empire."
-          },
-          {
-            "misconception": "Life in Britain returned to pre-Roman ways immediately",
-            "description": "Pupils might assume that Britain reverted quickly. Explain that Roman influences remained for many years."
-          }
-        ]
-      }
-    },
-    {
-      "type": "patch",
-      "reasoning": "The keywords are chosen to support pupils' understanding of key terms and concepts related to the lesson.",
-      "value": {
-        "type": "keywords",
-        "op": "add",
-        "path": "/keywords",
-        "value": [
-          {
-            "keyword": "Withdrawal",
-            "definition": "When a group or army leaves a place"
-          },
-          {
-            "keyword": "Anglo-Saxons",
-            "definition": "People who migrated to Britain after the Romans left"
-          },
-          {
-            "keyword": "Empire",
-            "definition": "A group of countries or regions controlled by one ruler or government"
-          }
-        ]
-      }
-    }
-  ],
-  "prompt": {
-    "type": "text",
-    "value": "Are you happy with the prior knowledge, key learning points, misconceptions, and keywords?\n\nIf not, select **'Retry'** or type an edit in the text box below.\n\nWhen you are happy with this section, tap **'Continue'** and I will suggest the content for your starter and exit quizzes and the learning cycles."
-  }
-}
