--- apps/nextjs/src/app/api/chat/fixtures/recordings/chat-3.chunks.txt
@@ -1,75 +1,0 @@
-{"
-type":"
-ll
-m
-Message
-","patches":[]
-,"prompt":{"
-type":"
-text
-","value":"
-I've
- completed
- the
- consistency
- check
-,
- and
- everything
- looks
- great
-!
- The
- lesson
- plan
- is
- consistent
-,
- uses
- British
- English
- correctly
-,
- and
- aligns
- with
- the
- learning
- objectives
-.
- You
- can
- now
- proceed
- to
- **
-create
- slides
-**,
- **
-download
- resources
-**,
- or
- **
-share
-**
- the
- lesson
- plan
-.
- If
- you
- have
- any
- further
- edits
- or
- questions
-,
- feel
- free
- to
- ask
-!"
-}}
