--- apps/nextjs/src/app/api/chat/fixtures/recordings/chat-4.chunks.txt
@@ -1,40 +1,0 @@
-{"
-type":"
-ll
-m
-Message
-","patches":[]
-,"prompt":{"
-type":"
-text
-","value":"
-If
- you
- need
- any
- more
- assistance
- or
- have
- additional
- questions
-,
- feel
- free
- to
- reach
- out
-.
- Good
- luck
- with
- your
- lesson
- on
- the
- end
- of
- Roman
- Britain
-!"
-}}
