--- apps/nextjs/src/app/api/chat/fixtures/recordings/chat-4.formatted.txt
@@ -1,8 +1,0 @@
-{
-  "type": "llmMessage",
-  "patches": [],
-  "prompt": {
-    "type": "text",
-    "value": "If you need any more assistance or have additional questions, feel free to reach out. Good luck with your lesson on the end of Roman Britain!"
-  }
-}
