+++ apps/nextjs/src/app/api/chat/fixtures/recordings/roman-britain-3-consistency.formatted.txt
@@ -1,0 +1,129 @@
+{
+  "type": "llmMessage",
+  "patches": [
+    {
+      "type": "patch",
+      "reasoning": "The starter quiz checks pupils' understanding of prior knowledge about Roman Britain before introducing new concepts.",
+      "value": {
+        "type": "quiz",
+        "op": "add",
+        "path": "/starterQuiz",
+        "value": [
+          {
+            "question": "When did the Romans invade Britain?",
+            "answers": [
+              "AD 43"
+            ],
+            "distractors": [
+              "AD 410",
+              "AD 1066"
+            ]
+          },
+          {
+            "question": "What did the Romans introduce to Britain?",
+            "answers": [
+              "Roads and villas"
+            ],
+            "distractors": [
+              "Castles and knights",
+              "Televisions and radios"
+            ]
+          },
+          {
+            "question": "How long did Roman Britain last?",
+            "answers": [
+              "Nearly 400 years"
+            ],
+            "distractors": [
+              "100 years",
+              "50 years"
+            ]
+          },
+          {
+            "question": "What was the role of the Roman army in Britain?",
+            "answers": [
+              "Protected from invasions"
+            ],
+            "distractors": [
+              "Built castles",
+              "Farmed the land"
+            ]
+          },
+          {
+            "question": "What technologies did Romans bring to Britain?",
+            "answers": [
+              "New technologies and ideas"
+            ],
+            "distractors": [
+              "Old technologies",
+              "Modern computers"
+            ]
+          },
+          {
+            "question": "What was a key feature of Roman society in Britain?",
+            "answers": [
+              "Towns and roads"
+            ],
+            "distractors": [
+              "Skyscrapers",
+              "Internet cafes"
+            ]
+          }
+        ]
+      }
+    },
+    {
+      "type": "patch",
+      "reasoning": "The first learning cycle focuses on identifying reasons for Roman Britain's decline, using explanations, checks, practice, and feedback.",
+      "value": {
+        "type": "cycle",
+        "op": "add",
+        "path": "/cycle1",
+        "value": {
+          "title": "Identifying Reasons for Decline",
+          "durationInMinutes": 15,
+          "explanation": {
+            "spokenExplanation": [
+              "Explain that Roman Britain's decline was due to economic troubles, including high taxes and reliance on slavery.",
+              "Discuss how Roman soldiers were needed back in Rome to protect against invaders, weakening Britain's defences.",
+              "Highlight the internal struggles within the Roman Empire that diverted attention from Britain.",
+              "Mention that Roman Britain faced local rebellions that strained resources.",
+              "Connect these issues to the broader context of the Roman Empire's instability."
+            ],
+            "accompanyingSlideDetails": "A simple timeline showing key events leading to the decline of Roman Britain.",
+            "imagePrompt": "Timeline of Roman Britain's decline",
+            "slideText": "Roman Britain's decline was due to economic troubles, military needs, and internal struggles in Rome."
+          },
+          "checkForUnderstanding": [
+            {
+              "question": "What was one reason for Roman Britain's decline?",
+              "answers": [
+                "Economic troubles"
+              ],
+              "distractors": [
+                "Peaceful expansion",
+                "Technological advances"
+              ]
+            },
+            {
+              "question": "Why were Roman soldiers needed back in Rome?",
+              "answers": [
+                "To protect against invaders"
+              ],
+              "distractors": [
+                "To build roads",
+                "To farm the land"
+              ]
+            }
+          ],
+          "practice": "List three reasons why Roman Britain began to decline. Write a sentence explaining each reason.",
+          "feedback": "Model answer: Economic troubles due to high taxes, military focus needed in Rome, and internal struggles within the Empire."
+        }
+      }
+    }
+  ],
+  "prompt": {
+    "type": "text",
+    "value": "The starter quiz and the first learning cycle have been added. Are you happy with the starter quiz and the content of the first learning cycle?\n\nIf not, select **'Retry'** or type an edit in the text box below.\n\nWhen you are happy with this section, tap **'Continue'** and I will generate the next learning cycles and the exit quiz."
+  }
+}
